update: crud event

+++ src/evenement/evenement.service.ts
@@ -2,10 +2,17 @@ import { ForbiddenException, Injectable } from '@nestjs/common';
 import { PrismaService } from '../prisma/prisma.service';
 import {
   CreateEvenementDto,
+  SwitchOnHeaderDto,
+  SwitchOnSubscribeDto,
   UpdateEvenementDto,
   UpdateIllustrationDto,
 } from './dto';
-import { CreateEvenement, GetEvenement } from './types';
+import {
+  CreateEvenement,
+  GetEvenement,
+  SwitchOnHeader,
+  SwitchOnSubscribe,
+} from './types';
 import * as fs from 'fs';
 
 @Injectable()
@@ -66,7 +73,7 @@ export class EvenementService {
         description: dto.description,
         illustration: dto.illustration,
         contenu: dto.contenu,
-        deadline: dto.deadline,
+        deadline: new Date(dto.deadline),
         onHeader: false,
         onSubscribe: dto.onSubscribe,
       },
@@ -75,6 +82,9 @@ export class EvenementService {
 
   async getEvenement(): Promise<GetEvenement[]> {
     const evenement = await this.prisma.evenement.findMany({
+      orderBy: {
+        id: 'desc',
+      },
       select: {
         id: true,
         idEntreprise: true,
@@ -97,7 +107,87 @@ export class EvenementService {
     });
 
     if (!evenement) {
-      throw new ForbiddenException("Il n'y a aucun article!");
+      throw new ForbiddenException("Il n'y a aucun évenement!");
+    }
+
+    return evenement;
+  }
+
+  async getEvenementAdmin(id: number): Promise<GetEvenement[]> {
+    const user = await this.prisma.utilisateur.findUnique({
+      where: {
+        id: Number(id),
+      },
+    });
+    if (!user) {
+      throw new ForbiddenException("L'utilisateur sélectionné n'éxiste pas");
+    } else if (user.role !== 1) {
+      const evenement = await this.prisma.evenement.findMany({
+        orderBy: {
+          id: 'desc',
+        },
+        where: {
+          entreprise: {
+            utilisateur: {
+              id: Number(id),
+            },
+          },
+        },
+        select: {
+          id: true,
+          idEntreprise: true,
+          titre: true,
+          slug: true,
+          illustration: true,
+          description: true,
+          contenu: true,
+          deadline: true,
+          onHeader: true,
+          createdAt: true,
+          onSubscribe: true,
+          inscription_evenement: {
+            select: {
+              id: true,
+              idUtilisateur: true,
+            },
+          },
+        },
+      });
+
+      if (!evenement) {
+        throw new ForbiddenException("Il n'y a aucun évenement!");
+      }
+
+      return evenement;
+    }
+
+    const evenement = await this.prisma.evenement.findMany({
+      orderBy: {
+        id: 'desc',
+      },
+      select: {
+        id: true,
+        idEntreprise: true,
+        titre: true,
+        slug: true,
+        illustration: true,
+        description: true,
+        contenu: true,
+        deadline: true,
+        onHeader: true,
+        createdAt: true,
+        onSubscribe: true,
+        inscription_evenement: {
+          select: {
+            id: true,
+            idUtilisateur: true,
+          },
+        },
+      },
+    });
+
+    if (!evenement) {
+      throw new ForbiddenException("Il n'y a aucun évenement!");
     }
 
     return evenement;
@@ -138,57 +228,93 @@ export class EvenementService {
 
   async updateEvenementBySlug(
     slug: string,
+    id: number,
     dto: UpdateEvenementDto,
   ): Promise<CreateEvenement> {
     const evenementExists = await this.prisma.evenement.findUnique({
       where: {
         slug,
       },
+      include: {
+        entreprise: {
+          include: {
+            utilisateur: true,
+          },
+        },
+      },
     });
 
     if (!evenementExists) {
       throw new ForbiddenException("Cet evenement n'existe pas!");
+    } else {
+      const updatedSlug = dto.titre
+        .toLocaleLowerCase()
+        .replace(/[&\/\\#,+()$~%.'":*?<>{}]/g, '')
+        .trim()
+        .split(' ')
+        .join('-');
+
+      const titreEvenementExist = await this.prisma.evenement.findUnique({
+        where: {
+          slug: updatedSlug,
+        },
+      });
+
+      if (titreEvenementExist && slug !== updatedSlug) {
+        throw new ForbiddenException("Ce titre d'évenement existe déja!");
+      }
+
+      const redacteur = await this.prisma.utilisateur.findUnique({
+        where: {
+          id,
+        },
+        include: {
+          role_utilisateur: {
+            select: {
+              id: true,
+            },
+          },
+        },
+      });
+
+      if (
+        evenementExists.entreprise.utilisateur.id !== id &&
+        redacteur?.role_utilisateur?.id !== 1
+      ) {
+        throw new ForbiddenException("Cet évènement n'est pas le votre!");
+      }
+
+      return await this.prisma.evenement.update({
+        data: {
+          titre: dto.titre,
+          description: dto.description,
+          contenu: dto.contenu,
+          deadline: new Date(dto.deadline),
+          slug: updatedSlug,
+          onSubscribe: dto.onSubscribe,
+        },
+        where: {
+          slug,
+        },
+      });
     }
-
-    const updatedSlug = dto.titre
-      .toLocaleLowerCase()
-      .replace(/[&\/\\#,+()$~%.'":*?<>{}]/g, '')
-      .trim()
-      .split(' ')
-      .join('-');
-
-    const titreEvenementExist = await this.prisma.evenement.findUnique({
-      where: {
-        slug: updatedSlug,
-      },
-    });
-
-    if (titreEvenementExist && slug !== updatedSlug) {
-      throw new ForbiddenException("Ce titre d'évenement existe déja!");
-    }
-
-    return await this.prisma.evenement.update({
-      data: {
-        titre: dto.titre,
-        description: dto.description,
-        contenu: dto.contenu,
-        deadline: dto.deadline,
-        slug: updatedSlug,
-        onSubscribe: dto.onSubscribe,
-      },
-      where: {
-        slug,
-      },
-    });
   }
 
   async updateIllustrationBySlug(
     slug: string,
+    id: number,
     dto: UpdateIllustrationDto,
   ): Promise<CreateEvenement> {
     const evenementBySlug = await this.prisma.evenement.findUnique({
       where: {
         slug,
+      },
+      include: {
+        entreprise: {
+          include: {
+            utilisateur: true,
+          },
+        },
       },
     });
 
@@ -202,6 +328,26 @@ export class EvenementService {
       fs.unlinkSync(`./images/${evenementBySlug.illustration}`);
     }
 
+    const redacteur = await this.prisma.utilisateur.findUnique({
+      where: {
+        id,
+      },
+      include: {
+        role_utilisateur: {
+          select: {
+            id: true,
+          },
+        },
+      },
+    });
+
+    if (
+      evenementBySlug.entreprise.idRedacteur !== id &&
+      redacteur?.role_utilisateur?.id !== 1
+    ) {
+      throw new ForbiddenException("Cet évènement n'est pas le votre!");
+    }
+
     return await this.prisma.evenement.update({
       data: {
         illustration: dto.illustration,
@@ -212,10 +358,20 @@ export class EvenementService {
     });
   }
 
-  async deleteEvenementBySlug(slug: string): Promise<CreateEvenement> {
+  async deleteEvenementBySlug(
+    slug: string,
+    id: number,
+  ): Promise<CreateEvenement> {
     const evenementBySlug = await this.prisma.evenement.findUnique({
       where: {
         slug,
+      },
+      include: {
+        entreprise: {
+          include: {
+            utilisateur: true,
+          },
+        },
       },
     });
 
@@ -227,9 +383,85 @@ export class EvenementService {
       fs.unlinkSync(`./images/${evenementBySlug.illustration}`);
     }
 
+    const redacteur = await this.prisma.utilisateur.findUnique({
+      where: {
+        id,
+      },
+      include: {
+        role_utilisateur: {
+          select: {
+            id: true,
+          },
+        },
+      },
+    });
+
+    if (
+      evenementBySlug.entreprise.idRedacteur !== id &&
+      redacteur?.role_utilisateur?.id !== 1
+    ) {
+      throw new ForbiddenException("Cet évènement n'est pas le votre!");
+    }
+
     return await this.prisma.evenement.delete({
       where: {
         slug,
+      },
+    });
+  }
+
+  async switchOnHeaderBySlug(
+    slug: string,
+    dto: SwitchOnHeaderDto,
+  ): Promise<SwitchOnHeader> {
+    const eventBySlug = await this.prisma.evenement.findUnique({
+      where: {
+        slug,
+      },
+    });
+
+    if (!eventBySlug) {
+      throw new ForbiddenException("Cet évenement n'existe pas!");
+    }
+
+    return await this.prisma.evenement.update({
+      data: {
+        onHeader: dto.onHeader,
+      },
+      where: {
+        slug,
+      },
+      select: {
+        id: true,
+        onHeader: true,
+      },
+    });
+  }
+
+  async switchOnSubscribeBySlug(
+    slug: string,
+    dto: SwitchOnSubscribeDto,
+  ): Promise<SwitchOnSubscribe> {
+    const eventBySlug = await this.prisma.evenement.findUnique({
+      where: {
+        slug,
+      },
+    });
+
+    if (!eventBySlug) {
+      throw new ForbiddenException("Cet évenement n'existe pas!");
+    }
+
+    return await this.prisma.evenement.update({
+      data: {
+        onSubscribe: dto.onSubscribe,
+      },
+      where: {
+        slug,
+      },
+      select: {
+        id: true,
+        onSubscribe: true,
       },
     });
   }
